backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "snapdrop-backend"
version = "0.1.0"
description = "SnapDrop backend API"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "sqlalchemy[asyncio]>=2.0.30",
    "asyncpg>=0.29.0",
    "alembic>=1.13.1",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.2.1",
    "python-jose[cryptography]>=3.3.0",
    "httpx>=0.27.0",
    "boto3>=1.34.0",
    "qrcode[pil]>=7.4.2",
    "Pillow>=10.3.0",
    "python-multipart>=0.0.9",
    "apscheduler>=3.10.4",
    "resend>=0.8.0",
    "slowapi>=0.1.9",
    "structlog>=24.1.0",
    "greenlet>=3.0.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.6",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "factory-boy>=3.3.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
